GlobalCommandsUpdate

+++ src/main/java/org/example/organization/Coordinates.java
@@ -5,10 +5,10 @@ package org.example.organization;
  */
 public class Coordinates {
     // Поле [ранящее значение x
-    private Float x; //Поле не может быть null
+    private Integer x; //Поле не может быть null
     private double y; //Максимальное значение поля: 274
 
-    public Coordinates(Float x, double y) {
+    public Coordinates(Integer x, double y) {
         this.x = x;
         this.y = y;
     }
